refractor Report and fileTools
