+++ composer.json
@@ -28,7 +28,7 @@
       "HexletPsrLinter\\": "src/"
     },
     "files": [
-      "src/fileTools.php",
+      "src/tools.php",
       "src/linters.php"
     ]
   }
