+++ src/Report/Format/ReportJson.php
@@ -1,10 +1,6 @@
 <?php
 
-namespace HexletPsrLinter\Report;
-
-use League\CLImate\CLImate;
-use League\CLImate\Util\Reader\ReaderInterface;
-use Symfony\Component\Yaml\Yaml;
+namespace HexletPsrLinter\Report\Format;
 
 /**
  * @property $cli CLImate
