+++ src/Report/Report.php
@@ -2,6 +2,9 @@
 
 namespace HexletPsrLinter\Report;
 
+use HexletPsrLinter\Report\Format\ReportJson;
+use HexletPsrLinter\Report\Format\ReportTxt;
+use HexletPsrLinter\Report\Format\ReportYaml;
 use League\CLImate\CLImate;
 
 /**
@@ -39,18 +42,9 @@ class Report
     }
 
     /**
-     * @param $file string
-     * @param $message Message
-     */
-    public function addLog($file, $message)
-    {
-        $this->logs[$file][] = $message;
-    }
-
-    /**
      * @param $report mixed
      */
-    public function loadReport($report)
+    private function loadReport($report)
     {
         foreach ($report as $log) {
             foreach ($log as $file => $message) {
@@ -60,6 +54,15 @@ class Report
                 $this->logs[$file] = $message;
             }
         }
+    }
+
+    /**
+     * @param $file string
+     * @param $message Message
+     */
+    public function addLog($file, $message)
+    {
+        $this->logs[$file][] = $message;
     }
 
     public function getLogs()
@@ -75,37 +78,5 @@ class Report
     public function printFormat()
     {
         return $this->formatReport->createReport($this->getLogs());
-    }
-
-    /**
-     * print report to console
-     */
-    public function printCli()
-    {
-        foreach ($this->getLogs() as $file => $messages) {
-            $this->cli->lightBlue()->bold()->inline($file)->br();
-            /** @var $message Message */
-            foreach ($messages as $message) {
-                $this->cli->white()->bold()->inline(sprintf('%-5s', $message->getLine()));
-
-                $format = '%-10s';
-                $text = $message->getLevel();
-                switch ($text) {
-                    case self::LOG_LEVEL_ERROR:
-                        $this->cli->red()->inline(sprintf($format, $text));
-                        break;
-                    case self::LOG_LEVEL_WARNING:
-                        $this->cli->yellow()->inline(sprintf($format, $text));
-                        break;
-                    case self::LOG_LEVEL_FIXED:
-                        $this->cli->green()->inline(sprintf($format, $text));
-                        break;
-                }
-
-                $this->cli->lightCyan()->bold()->inline(sprintf('%-25s', $message->getName()));
-                $this->cli->white()->inline($message->getMessage())->br();
-            }
-        }
-        $this->cli->br();
     }
 }
